+++ src/components/index.ts
@@ -17,7 +17,7 @@ export { default as SyTextField } from './Customs/SyTextField/SyTextField.vue'
 export { default as DataList } from './DataList/DataList.vue'
 export { default as DataListGroup } from './DataListGroup/DataListGroup.vue'
 export { default as DataListItem } from './DataListItem/DataListItem.vue'
-export { default as DatePickerInput } from './DatePickerInput/DatePickerInput.vue'
+export { default as DatePicker } from '@/components/DatePicker/DatePicker.vue'
 export { default as DialogBox } from './DialogBox/DialogBox.vue'
 export { default as DownloadBtn } from './DownloadBtn/DownloadBtn.vue'
 export { default as ErrorPage } from './ErrorPage/ErrorPage.vue'
